Add shutdown test to cloud tests
